pyproject.toml
[project]
name = "moy-sklad-api"
version = "0.3"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "aiohttp>=3.13.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tzdata>=2025.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ms-api"
Documentation = "https://github.com/yourusername/ms-api#readme"
Repository = "https://github.com/yourusername/ms-api"
Issues = "https://github.com/yourusername/ms-api/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["moy_sklad_api"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

